requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.1
django-admin==1.3.2
django-excel-response2==2.0.8
django-model-utils==3.1.2
django-phonenumber-field==2.4.0
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.9.4
idna==2.8
mysqlclient==1.4.2.post1
phonenumbers==8.10.12
pycparser==2.19
PyJWT==1.7.1
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
screen==1.0.1
six==1.12.0
urllib3==1.25.3
vine==1.3.0
xlwt==1.3.0
